Refactor: Move test account data imports from universal to server directory and remove deprecated auth.development.ts
